partitions basic grading program
